+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# [1.4.0](https://github.com/BlockABC/abcwallet.js/compare/v1.3.0...v1.4.0) (2020-03-24)
+
+
+### Bug Fixes
+
+* fix api field ([9adddc1](https://github.com/BlockABC/abcwallet.js/commit/9adddc149782ff1c88a839bb7f08f4a295f2c51a))
+* fix version bug ([f2455bd](https://github.com/BlockABC/abcwallet.js/commit/f2455bddf92033a4a84206095bb615fc1cde5e01))
+
+
+### Features
+
+* add abcwallet client detection api ([9977d7b](https://github.com/BlockABC/abcwallet.js/commit/9977d7b5909669d99eeb548a4347c96fd1e950af))
+* add browser api ([328217d](https://github.com/BlockABC/abcwallet.js/commit/328217df03e69ec14bfe8452907963a84242cd91))
+* add compareVersion ([f271768](https://github.com/BlockABC/abcwallet.js/commit/f2717688dba88c72623de6d1295e0be241cf1dbd))
+* add tools to invoke client ([207ee58](https://github.com/BlockABC/abcwallet.js/commit/207ee586bfba0ef0cc2ab5b2d37bf7d12dedeac7))
+* use dynamic version from package.json ([f888fe8](https://github.com/BlockABC/abcwallet.js/commit/f888fe8b0c6afd56d8dd60b97a1f8876a94dc98c))
+
 # [1.3.0](https://github.com/BlockABC/abcwallet.js/compare/v1.2.0...v1.3.0) (2020-03-09)